Work in progress on floating point editable height maps

+++ WorldPainter/WPCore/src/main/java/org/pepsoft/worldpainter/heightMaps/MaximisingHeightMap.java
@@ -22,7 +22,7 @@ public class MaximisingHeightMap extends CombiningHeightMap {
     }
 
     @Override
-    public float getHeight(int x, int y) {
+    public float getHeight(float x, float y) {
         return Math.max(heightMap1.getHeight(x, y), heightMap2.getHeight(x, y));
     }
