+++ WorldPainter/WPCore/src/main/java/org/pepsoft/worldpainter/heightMaps/ProductHeightMap.java
@@ -21,7 +21,7 @@ public class ProductHeightMap extends CombiningHeightMap {
     }
 
     @Override
-    public float getHeight(int x, int y) {
+    public float getHeight(float x, float y) {
         return heightMap1.getHeight(x, y) * heightMap2.getHeight(x, y);
     }
